修复已知bug

+++ app/src/main/java/com/specher/superhookbox/Config.java
@@ -24,32 +24,43 @@ import java.lang.reflect.Method;
 import java.util.Iterator;
 import java.util.List;
 
-public class Config extends FileProvider {
-    public File rootPath = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOCUMENTS);
+public class Config {
+    private final File rootPath = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOCUMENTS);
     private  int count=0;
-    public final String tiktokShow = "支持版本:15.X以及15.X精简版";
-    public final String telegramShow = "重定向Telegram缓存文件夹到系统Pictures文件夹，开启后请手动移动SD卡目录下的Telegram文件夹到Pictures中。\n" +
+    private static final String tiktokShow = "支持版本:15.X以及15.X精简版";
+    private static final String telegramShow = "重定向Telegram缓存文件夹到系统Pictures文件夹，开启后请手动移动SD卡目录下的Telegram文件夹到Pictures中。\n" +
             "开启删除.nomedia可以让文件夹在相册中出现，方便同步到云端。\n" +
             "阻止删除消息重新打开聊天界面删除的消息即会出现。\n" +
             "支持版本：Telegram官方版/Beta版/科学版/TG Plus/Nekogram/Nekogram X";
-    private final String jsonFilename;
+    private static final String configName_tiktok = "tiktok.json";
+    private static final String configName_tg = "Telegram.json";
+    private static final String configName_hookbox = "HookBox.json";
+    public String jsonFilename;
     public JSONObject globalJSON = new JSONObject();
-    public String isTikTok = "开启抖X功能";
+    public static String isTikTok = "开启抖X功能";
     public String isAutoPlay = "自动播放下一条";
     public String downLoadVideo = "无水印下载";
+    public String isFullScreen = "去除视频黑边";
+    public String isJumpSplashAd = "跳过启动广告";
+    public String isCommentDark = "评论暗黑模式";
     public String jumpAD = "跳过视频广告";
-    public String hideRightMenu = "隐藏右侧按钮和文字";
+    public String jumpADTip = "跳过广告时提示";
+    public String hideRightMenu = "全屏时隐藏右侧按钮";
+    public String hideAwemeIntro = "全屏时隐藏文字";
     public String fullVideoPlay = "长按切换全屏模式";
     public String hideStatusBar = "全屏时隐藏状态栏";
     public String hideBottomTab = "全屏时隐藏底栏";
     public String hideTopTab = "全屏时隐藏顶栏";
-    public String isTelegram = "开启Telegram功能";
+    public static String isTelegram = "开启Telegram功能";
     public String storageRedirect = "重定向存储";
     public String delNomedia = "删除.nomedia文件";
     public String unRecalled = "阻止删除消息";
     public String unDelete = "阻止消息自毁(阅后即焚)";
+    public static String isHookBox = "开启HookBox";
     public String isFirst = "首次启动"+BuildConfig.VERSION_CODE;
     private final Context mContext;
+
+
 
     Config(Context context, String jsonFilename) throws Exception {
         this.mContext = context;
@@ -57,11 +68,30 @@ public class Config extends FileProvider {
         initPref();
     }
 
+    public static String getConfigName(String optionName){
+        if(optionName.equals(isTikTok)){
+            return configName_tiktok;
+        }else if(optionName.equals(isTelegram)){
+            return  configName_tg;
+        }else if(optionName.equals(isHookBox)) {
+            return configName_hookbox;
+        }else{
+            return null;
+        }
+    }
+    public String getShow(String optionName){
+        if(optionName.equals(isTikTok)){
+            return tiktokShow;
+        }if(optionName.equals(isTelegram)){
+            return  telegramShow;
+        }else return null;
+    }
+
     public JSONObject readPref() throws Exception {
         InputStreamReader fileReader = new InputStreamReader(new FileInputStream(new File(rootPath, jsonFilename)));
         BufferedReader fp = new BufferedReader(fileReader);
         String result = "";
-        String re = "";
+        String re;
         while (true) {
             String readLine = fp.readLine();
             re = readLine;
@@ -115,25 +145,30 @@ public class Config extends FileProvider {
           }
         }
         switch (jsonFilename) {
-            case "tiktok.json":
+            case configName_tiktok:
                 this.globalJSON.put(this.isAutoPlay, false);
+                this.globalJSON.put(this.isFullScreen, true);
+                this.globalJSON.put(this.isCommentDark, false);
                 this.globalJSON.put(this.downLoadVideo, false);
+                this.globalJSON.put(this.isJumpSplashAd, false);
                 this.globalJSON.put(this.jumpAD, false);
+                this.globalJSON.put(this.jumpADTip, false);
                 this.globalJSON.put(this.fullVideoPlay, false);
                 this.globalJSON.put(this.hideStatusBar, false);
                 this.globalJSON.put(this.hideBottomTab, true);
                 this.globalJSON.put(this.hideTopTab, true);
                 this.globalJSON.put(this.hideRightMenu, true);
+                this.globalJSON.put(this.hideAwemeIntro, false);
                 break;
-            case "Telegram.json":
+            case configName_tg:
                 this.globalJSON.put(this.storageRedirect, false);
                 this.globalJSON.put(this.delNomedia, false);
                 this.globalJSON.put(this.unRecalled, false);
                 this.globalJSON.put(this.unDelete, false);
                 break;
-            case "HookBox.json":
-                this.globalJSON.put(this.isTikTok, false);
-                this.globalJSON.put(this.isTelegram, false);
+            case configName_hookbox:
+                this.globalJSON.put(isTikTok, false);
+                this.globalJSON.put(isTelegram, false);
                 this.globalJSON.put(this.isFirst, true);
 
                 break;
@@ -149,13 +184,21 @@ public class Config extends FileProvider {
         try{
         while (keys.hasNext()) {
             if (!readPref().has(keys.next())) {
-                jsonFile.delete();
+                //配置文件有变化或者错误
+                count++;
+                if(count>2){
+                    Utils.exitApp(mContext);
+                }else{
+                    Toast.makeText(mContext,"配置文件有更新，请重新配置。",Toast.LENGTH_SHORT).show();
+                    jsonFile.delete();
+                    initPref();
+                }
             }
         }
         }
         catch(Exception e){
             Log.i("HookBox",e.getMessage());
-            Toast.makeText(mContext,"配置文件损坏，需要重新配置。"+e.toString(),Toast.LENGTH_LONG).show();
+            Toast.makeText(mContext,"读取配置异常，请检查是否给与全部文件读写权限。"+e.toString(),Toast.LENGTH_LONG).show();
             count++;
             if(count>2){
                 Utils.exitApp(mContext);
@@ -165,6 +208,4 @@ public class Config extends FileProvider {
             }
         }
     }
-
-
 }
